requirements.txt
google-generativeai
python-dotenv
pyttsx3
SpeechRecognition
PyAudio
pyautogui
wikipedia
requests
langdetect
pillow
